app/requirements.txt
cassandra-driver
pathvalidate
venv-pack
